feat: Add fan_only mode to HVAC controls and toggle light script
This commit adds the "fan_only" mode to the HVAC controls in the lovelace.dashboard_unknown file. It also includes a new script called "toggle_light" in the scripts.yaml file. These changes enhance the functionality of the home automation system by allowing control over the fan mode of the HVAC system and providing a script to toggle the light.